+++ app/src/main/java/com/example/weatheralarmapp/domain/usecase/alarm/AddAlarmItemUseCase.kt
@@ -1,4 +1,4 @@
-package com.example.weatheralarmapp.domain.usecase
+package com.example.weatheralarmapp.domain.usecase.alarm
 
 import android.app.AlarmManager
 import com.example.weatheralarmapp.data.local.AlarmItem
@@ -6,20 +6,17 @@ import com.example.weatheralarmapp.data.repository.AlarmItemRepository
 import com.example.weatheralarmapp.domain.AlarmHelper
 import javax.inject.Inject
 
-class DeleteAlarmItemUseCase
+class AddAlarmItemUseCase
     @Inject
     constructor(
         private val alarmHelper: AlarmHelper,
         private val alarmItemRepository: AlarmItemRepository,
     ) {
-        suspend fun deleteAlarmItem(
-            alarmItem: AlarmItem,
+        suspend fun addAlarmItem(
             alarmManager: AlarmManager,
+            alarmItem: AlarmItem,
         ) {
-            alarmHelper.cancelAlarm(
-                alarmManager,
-                alarmItem,
-            )
-            alarmItemRepository.deleteAlarmItem(alarmItem)
+            alarmItemRepository.insertAlarmItem(alarmItem)
+            alarmHelper.setAlarm(alarmManager, alarmItem, false)
         }
     }
