+++ app/src/main/java/com/example/weatheralarmapp/domain/usecase/alarm/DeleteAlarmItemUseCase.kt
@@ -1,4 +1,4 @@
-package com.example.weatheralarmapp.domain.usecase
+package com.example.weatheralarmapp.domain.usecase.alarm
 
 import android.app.AlarmManager
 import com.example.weatheralarmapp.data.local.AlarmItem
@@ -6,17 +6,20 @@ import com.example.weatheralarmapp.data.repository.AlarmItemRepository
 import com.example.weatheralarmapp.domain.AlarmHelper
 import javax.inject.Inject
 
-class AddAlarmItemUseCase
+class DeleteAlarmItemUseCase
     @Inject
     constructor(
         private val alarmHelper: AlarmHelper,
         private val alarmItemRepository: AlarmItemRepository,
     ) {
-        suspend fun addAlarmItem(
-            alarmManager: AlarmManager,
+        suspend fun deleteAlarmItem(
             alarmItem: AlarmItem,
+            alarmManager: AlarmManager,
         ) {
-            alarmItemRepository.insertAlarmItem(alarmItem)
-            alarmHelper.setAlarm(alarmManager, alarmItem, false)
+            alarmHelper.cancelAlarm(
+                alarmManager,
+                alarmItem,
+            )
+            alarmItemRepository.deleteAlarmItem(alarmItem)
         }
     }
